requirements.txt
Django==2.2.13
django-crispy-forms==1.7.2
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
Pillow==7.2.0
PyJWT==1.7.1
pytz==2018.9
sqlparse==0.3.1
